bug-fix (see desc)
-app no loger crashes if Documents/BoA does not have a .md file; \n-correctly delates old indexes of updated files; \n-recursive index for all files that are in subfolders of Documents/BoA

+++ model/IndexManager.py
@@ -43,14 +43,18 @@ class IndexManager:
         """Re-embedds ALL the files and creates a new index"""
         existing_ids = self.chroma_collection.get()["ids"]
         if existing_ids:
-            # print(
-            #    f"Clearing {len(existing_ids)} old entries from ChromaDB collection..."
-            # )
+            print(
+                f"Clearing {len(existing_ids)} old entries from ChromaDB collection..."
+            )
             self.chroma_collection.delete(ids=existing_ids)
 
         documents = SimpleDirectoryReader(
-            self.docs_dir, required_exts=[".md"]
+            self.docs_dir, required_exts=[".md"], recursive=True
         ).load_data()
+
+        if not documents:
+            print("no ducuments were found")
+            return None
 
         storage_context = StorageContext.from_defaults(vector_store=self.vector_store)
 
@@ -78,22 +82,27 @@ class IndexManager:
             print(f"CromaDB not found at {CHROMA_DB_PATH}.")
             # should never really happen
             # we are creating a db at startup if not found
-            pass
         else:
-            storage_context = StorageContext.from_defaults(
-                vector_store=self.vector_store, persist_dir=PERSIST_DIR
-            )
+            # storage_context = StorageContext.from_defaults(
+            #    vector_store=self.vector_store, persist_dir=PERSIST_DIR
+            # )
 
-            index = load_index_from_storage(storage_context=storage_context)
-
-            retriever_engine = index.as_retriever()
+            # index = load_index_from_storage(storage_context=storage_context)
+            if not self.index:
+                self.re_build_all()
+            retriever_engine = self.index.as_retriever()
             response = retriever_engine.retrieve(query_msg)
 
             return sorted(response, key=lambda x: x.score, reverse=True)[0].text
 
     def delete_file_node(self, filepath: str) -> None:
         """Deletes all nodes associated with the provided file"""
-        self.chroma_collection.delete(where={"file_path": filepath})
+        if not os.path.exists(PERSIST_DIR):
+            print(f"CromaDB not found at {CHROMA_DB_PATH}.")
+            # should never really happen
+            # we are creating a db at startup if not found
+        else:
+            self.chroma_collection.delete(where={"file_path": filepath})
 
     def update_file_node(self, filepath: str) -> None:
         """
@@ -104,7 +113,8 @@ class IndexManager:
             print(f"No Index found at not found at {PERSIST_DIR}.")
             pass
         else:
-            self.index.delete_ref_doc(ref_doc_id=filepath, delete_from_docstore=True)
+            # self.index.delete_ref_doc(ref_doc_id=filepath, delete_from_docstore=True)
+            self.chroma_collection.delete(where={"file_path": filepath})
             reader = SimpleDirectoryReader(
                 input_dir=self.docs_dir,
                 input_files=[filepath],
@@ -117,4 +127,6 @@ class IndexManager:
             )
             new_nodes = node_parser.get_nodes_from_documents(documents)
 
+            if not self.index:
+                self.re_build_all()
             self.index.insert_nodes(new_nodes)
